refactoring of stepdefs and pages
