+++ src/test/scala/uk/gov/hmrc/test/ui/pages/CommonPage.scala
@@ -28,12 +28,6 @@ object CommonPage extends BasePage {
   val registrationUrl: String = TestConfiguration.url("ioss-registration-frontend")
   var credId: String          = "1234123412341234"
 
-  def goToRegistrationJourney(): Unit =
-    driver.navigate().to(registrationUrl)
-
-  def checkJourneyUrl(): Unit =
-    driver.getCurrentUrl shouldBe registrationUrl + "/ioss-registered"
-
   def checkUrl(url: String): Unit =
     driver.getCurrentUrl should startWith(s"${TestConfiguration.url("ioss-registration-frontend")}/$url")
 
@@ -48,13 +42,6 @@ object CommonPage extends BasePage {
 
   def clickContinue(): Unit =
     driver.findElement(By.id("continue")).click()
-
-  def selectChoice(data: String): Unit =
-    data match {
-      case "Yes"                    => driver.findElement(By.id("value_0")).click()
-      case "Yes, details incorrect" => driver.findElement(By.id("value_1")).click()
-      case "No, different business" => driver.findElement(By.id("value_2")).click()
-    }
 
   def enterData(data: String, inputId: String = "value"): Unit = {
     driver.findElement(By.id(inputId)).sendKeys(data)
@@ -78,7 +65,7 @@ object CommonPage extends BasePage {
     clickContinue()
   }
 
-  def waitForElement(by: By) =
+  def waitForElement(by: By): Unit =
     new FluentWait(driver).until(ExpectedConditions.presenceOfElementLocated(by))
 
   def selectValueAutocomplete(country: String): Unit = {
@@ -98,69 +85,14 @@ object CommonPage extends BasePage {
     CommonPage.clickContinue()
   }
 
-  def goToEmailVerificationPasscodeGeneratorUrl(): Unit =
-    driver.navigate
-      .to(
-        "http://localhost:10190/pay-vat-on-goods-sold-to-eu/register-for-import-one-stop-shop/test-only/get-passcodes"
-      )
-
-  def goToEmailVerificationUrl(journeyId: String, mode: String): Unit = {
-    val url = mode match {
-      case "registration"                => "bank-details"
-      case "change answers"              => "check-your-answers"
-      case "amend registration"          => "change-your-registration"
-      case "amend previous registration" => "change-a-previous-registration"
-      case "rejoin registration"         => "rejoin-registration"
-      case _                             =>
-        throw new Exception("URL doesn't exist")
-    }
-    driver
-      .navigate()
-      .to(
-        s"http://localhost:9890/email-verification/journey/$journeyId/passcode?continueUrl=/pay-vat-on-goods-sold-to-eu/register-for-import-one-stop-shop/$url&origin=IOSS&service=ioss-registration-frontend"
-      )
-  }
-
-  def enterPasscode(passcode: String): Unit = {
-    driver.findElement(By.id("passcode")).sendKeys(passcode)
-    driver.findElement(By.className("govuk-button")).click()
-  }
-
-  def checkBTA(): Unit =
-    driver.getCurrentUrl should endWith("business-account")
-
   def checkReturnsDashboard(): Unit =
     driver.getCurrentUrl should startWith(TestConfiguration.url("ioss-returns-frontend") + "/your-account")
-
-  def goToContinueOnSignInPage(): Unit =
-    driver
-      .navigate()
-      .to("http://localhost:10190/pay-vat-on-goods-sold-to-eu/register-for-import-one-stop-shop/continue-on-sign-in")
-
-  def selectContinueRegistration(data: String): Unit = {
-    data match {
-      case "Yes"                                  => driver.findElement(By.id("continueProgress")).click()
-      case "No,delete my answers and start again" => driver.findElement(By.id("deleteProgress")).click()
-      case _                                      => throw new Exception("Option doesn't exist")
-    }
-    CommonPage.clickContinue()
-  }
 
   def retrieveCredId(): String =
     credId
 
   def generateCredId(): Unit =
     credId = Random.between(1000000000000000L, 9000000000000000L).toString
-
-  def checkInterceptPage(): Unit =
-    driver.getCurrentUrl should startWith(TestConfiguration.url("ioss-returns-frontend") + "/intercept-unusable-email")
-
-  def checkBusinessContactDetails(): Unit =
-    driver.getCurrentUrl should startWith(
-      TestConfiguration.url(
-        "ioss-registration-frontend"
-      ) + "/business-contact-details?waypoints=change-your-registration"
-    )
 
   def clickBackButton(): Unit =
     driver
